finish create

+++ app/Http/Controllers/ArticleController.php
@@ -14,7 +14,7 @@ class ArticleController extends Controller
      */
     public function index()
     {
-        $articles = Article::all();
+        $articles = Article::orderBy('created_at', 'desc')->paginate(10);
         $data = ['articles' => $articles];
         return view('articles.index', $data);
     }
@@ -47,6 +47,7 @@ class ArticleController extends Controller
             'body' => 'required'
         ]);
         $article = new Article();
+        $article->user_id = \Auth::id();
         $article->title = $request->title;
         $article->body = $request->body;
         $article->save();
@@ -75,6 +76,7 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
+        $this->authorize($article);
         $data = ['article' => $article];
         return view('articles.edit', $data);
     }
@@ -88,6 +90,7 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
+        $this->authorize($article);
         $this->validate($request, [
             'title' => 'required|max:255',
             'body' => 'required'
@@ -106,7 +109,17 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        $this->authorize($article);
         $article->delete();
         return redirect(route('articles.index'));
+    }
+
+    public function bookmark_articles()
+    {
+        $articles = \Auth::user()->bookmark_articles()->orderBy('created_at', 'desc')->paginate(10);
+        $data = [
+            'articles' => $articles,
+        ];
+        return view('articles.bookmarks', $data);
     }
 }
